client/requirements.txt
# Client Dependencies
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pytest==7.4.3
torch==2.1.1
